remove allFileItems.
Treating deleted files as identical to their pre-deletion state is a bit excessive. The mechanism to reference files used for task management at a later time is good in itself. Viewing past tasks in a calendar view is an interesting attempt, but making the entire codebase more complex for this purpose has overwhelmingly more disadvantages.